requirements.txt
asgiref>=3.4.1
async-generator>=1.10
attrs>=21.2.0
boto3>=1.18.64
botocore>=1.21.64
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.7
colorama>=0.4.4
configparser>=5.0.2
coreapi>=2.3.3
coreschema>=0.0.4
crayons>=0.4.0
cryptography>=35.0.0
Django>=3.2.8
django-rest-framework>=0.1.0
djangorestframework>=3.12.4
drf-yasg>=1.20.0
h11>=0.12.0
idna>=3.3
inflection>=0.5.1
itypes>=1.2.0
Jinja2>=3.0.2
jmespath>=0.10.0
MarkupSafe>=2.0.1
outcome>=1.1.0
packaging>=21.0
pycparser>=2.20
pyOpenSSL>=21.0.0
pyparsing>=2.4.7
python-dateutil>=2.8.2
python-dotenv>=0.19.1
pytz>=2021.3
requests>=2.26.0
ruamel.yaml>=0.17.16
ruamel.yaml.clib>=0.2.6
s3transfer>=0.5.0
selenium>=4.0.0
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
sqlparse>=0.4.2
trio>=0.19.0
trio-websocket>=0.9.2
uritemplate>=4.1.1
urllib3>=1.26.7
webdriver-manager>=3.5.1
wsproto>=1.0.0